# btf: optimize stringTable to speed up vmlinux parsing